fix: set speed int->double fix

+++ oun-common/src/main/java/com/simol/ouncommon/healthset/dto/HealthSetCreateRequest.java
@@ -43,7 +43,7 @@ public class HealthSetCreateRequest {
 
     @Schema(description = "세트 속도(km/h)", example = "0")
     @JsonProperty("set_speed")
-    private int setSpeed;
+    private double setSpeed;
     
     @Schema(description = "설명", example = "스쿼트 설명")
     private String description;
